Lazy load the Kubernetes client context.
If we require the Kubernetes context to be loaded by an init function,
not even the help text can be shown without the user having a
current Kubernetes config. Switch over to lazily loading the cluster
configuration, which results in a similar API, but avoids the problems
around init functions.

Signed-off-by: James Peach <jpeach@vmware.com>