Bundle and apply YAML from the conformance command.
The conformance verification tests depend on specific versions
of resources being deployed into the cluster under tests. To make
distribution easier and to ensure that the checks and resources are always
in sync, embed the deployment YAML in the conformance tool and add a new
subcommand to run the tests. The new "apply" subcommand also allows us to
give the operator a way to specify the Ingress class.

This updates #11.
This updates #12.

Signed-off-by: James Peach <jpeach@vmware.com>